pyproject.toml
[project]
name = "app"
version = "1.0"
dependencies = [
    "torch>=2.4",
    "torchvision",
    "polars",
    "lightning>=2.4",
    "typer",
    "datasets[vision]",
    "hydra-core",
    "torchmetrics",
    "networkx",
    "wandb",
    "timm>=1.0.9",
]
requires-python = ">=3.12"
authors = [{ name = "Malte Sönnichsen" }]
readme = "README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "jupyterlab",
    "ipython",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.2",
    "scikit-learn",
    "altair[all]>=5.4.0",
    "scipy>=1.14.1",
]
